+++ yrs/src/block.rs
@@ -7,7 +7,7 @@ use crate::slice::{BlockSlice, GCSlice, ItemSlice};
 use crate::store::Store;
 use crate::transaction::TransactionMut;
 use crate::types::text::update_current_attributes;
-use crate::types::{Attrs, TypePtr, TypeRef, Value};
+use crate::types::{Attrs, TypePtr, TypeRef};
 use crate::undo::UndoStack;
 use crate::updates::decoder::{Decode, Decoder};
 use crate::updates::encoder::{Encode, Encoder};
