+++ yrs/src/block_iter.rs
@@ -2,8 +2,8 @@ use crate::block::{Item, ItemContent, ItemPtr, Prelim};
 use crate::branch::BranchPtr;
 use crate::moving::{Move, StickyIndex};
 use crate::transaction::{ReadTxn, TransactionMut};
-use crate::types::{TypePtr, Value};
-use crate::{Assoc, ID};
+use crate::types::TypePtr;
+use crate::{Assoc, Value, ID};
 
 /// Struct used for iterating over the sequence of item's values with respect to a potential
 /// [Move] markers that may change their order.
