+++ yrs/src/lib.rs
@@ -587,6 +587,7 @@ mod gc;
 pub mod iter;
 mod moving;
 pub mod observer;
+mod out;
 mod slice;
 mod state_vector;
 pub mod sync;
@@ -618,6 +619,7 @@ pub use crate::moving::IndexedSequence;
 pub use crate::moving::Offset;
 pub use crate::moving::StickyIndex;
 pub use crate::observer::{Observer, Subscription};
+pub use crate::out::Value;
 pub use crate::state_vector::Snapshot;
 pub use crate::state_vector::StateVector;
 pub use crate::store::Store;
@@ -653,7 +655,6 @@ pub use crate::types::GetString;
 pub use crate::types::Observable;
 pub use crate::types::RootRef;
 pub use crate::types::SharedRef;
-pub use crate::types::Value;
 pub use crate::update::Update;
 
 pub type UndoManager = crate::undo::UndoManager<()>;
